requirements.txt
scapy
drain3
pillow
numpy
# pytorch去官网下
torch
torchvision
tqdm
matplotlib
timm
transformers